add filter to traces page that shows/hides traces based on date range
