+++ client/src/components/TraceList.jsx
@@ -1,60 +1,14 @@
+import React, { useState, useEffect }  from "react"
 import DebugTraceDisplay from './DebugTraceDisplay'
 import spinner from '../assets/pulse-1.1s-200px.svg';
 import './HomeDisplay.css';
-import errorImage from '../assets/error-svgrepo-com.svg';
+import TraceSelector from './TraceSelector'
 import TraceFilters from './TraceFilters';
 
-
-const getFromRight = (s) => {
-  let result = '';
-  for (let i = s.length-1; i >= 0; i--) {
-    if (s[i] == '/') return result;
-    result = s[i] + result;
-  }
-  return result;
-}
-
-const findErrorsInTrace = (trace) => {
-  let errors = 0;
-
-  const process = (node) => {
-    if (node.fullData && node.fullData.Document.error) errors++;
-    if (node.children) node.children.forEach((n) => {process(n)})
-  }
-
-  process(trace);
-  return errors;
-}
-
-
 function TraceList (props) {
-    
-    const traces = [];
-    for (let n = 0; n < props.traces.length; n++) {
-        const url = props.traces[n].fullData.Document.http.request.url;
-        const endpt = getFromRight(url)
-        let errors = findErrorsInTrace(props.traces[n]);
-        traces.push(<button key={'tb'+Math.random()} 
-                            onClick={() => props.setCurrentTrace(n)}
-                            style={{fontSize: 'small'}}>
-                            <span>{endpt + ' - '
-                                }
-                            </span>
-                            <span style={{fontSize: 'x-small'}}>
-                             {props.traces[n].id}
-                            </span>
-                            <span style={{marginLeft: '4px'}}>
-                                {
-                                    (errors ? (<img src={errorImage} width='16px'></img>) : '')
-                                }
-                            </span>
-                            <span style={{color: 'white', fontSize: 'x-small'}}>
-                                {
-                                    (errors ? (' (' + errors + ' errors )') : '')
-                                }
-                            </span>
-                        </button>)
-    }
+
+  const [start_value, onChangeStart] = useState(new Date()-1000*60*60*24*7);
+  const [end_value, onChangeEnd] = useState(new Date());
 
   function refreshData() {
     //clears Traces table from redis
@@ -69,16 +23,19 @@ function TraceList (props) {
 
   return (
     <div className='trace-list-container'>
-      <div className='trace-list-traces'>
-        <p>TraceList</p>
+      <div >
         {props.loading && (
           <img className='loading-spinner' src={spinner} alt='Loading' />
         )}
-        <button onClick={() => refreshData()}>Refresh Data</button>
-        {traces}
       </div>
+      <TraceSelector traces={props.traces}
+                start_value={start_value}
+                end_value={end_value} 
+                setCurrentTrace = {props.setCurrentTrace} />
       <div style={{flexDirection: 'column', width: '100%'}}>
-        <TraceFilters handleRefreshData={refreshData} />
+        <TraceFilters handleRefreshData={refreshData} 
+              start_value={start_value} onChangeStart={onChangeStart}
+              end_value={end_value} onChangeEnd={onChangeEnd}/>
         <DebugTraceDisplay trace={props.traces[props.currentTrace]} />
       </div>
     </div>
